requirements.txt
langchain>=0.3.0
langchain-openai>=0.3.0
langgraph>=0.2.0
langchain-core>=0.3.0
openai>=1.50.0
click>=8.1.0
rich>=13.7.0
python-dotenv>=1.0.0
pydantic>=2.0.0
